Chrome to Chromium port
-Chrome* commands now run for Chrome, Edge, and Brave
-Removed HuntLOLBAS from running with -group=all
